% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{force_all}
\alias{force_all}
\title{Force all}
\usage{
force_all(...)
}
\arguments{
\item{...}{Parameters to change to a list.}
}
\description{
This is a helper function for ipea_pal() and is not exported.
}
\examples{
\dontrun{ if (interactive()) {
force_all <- function(...) list(...)
}}
}
