% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_text.R
\name{insert_text}
\alias{insert_text}
\title{Insert Text into Plot}
\usage{
insert_text(
  label = NULL,
  decimals = 0,
  show_percents = FALSE,
  vertical = TRUE,
  ...
)
}
\arguments{
\item{label}{The label variable indicating the text to be displayed.}

\item{decimals}{The number of decimal places to round the label values.}

\item{show_percents}{Logical.}

\item{vertical}{Logical.}

\item{...}{Additional arguments to be passed to the \code{geom_text} function from the \code{ggplot2} package.}
}
\value{
A \code{geom_text} layer with text labels added to the plot.
}
\description{
This function adds text labels to a plot at specified positions.
}
\examples{
\dontrun{ if (interactive()) {
library(ggplot2)
library(ipeaplot)

graph <- subset(abjData::pnud_uf, ano == 2010)

# Standard plot adding percentage notation
graph$perc <- (graph$rdpc/sum(graph$rdpc))*100
ggplot(graph, aes(x = ufn, y = perc))+
  geom_bar(stat = "identity", fill = 'blue', width= 0.5) +
  insert_text(label = "perc", show_percents = T)

# Default plot and changing to horizontal
plot <- ggplot(graph, aes(x = ufn, y = rdpc))+
  geom_bar(stat = "identity", fill = 'blue', width= 0.5) +
  insert_text(label = "rdpc", vertical = F)

# Default plot
plot <- ggplot(graph, aes(x = ufn, y = rdpc))+
  geom_bar(stat = "identity", fill = 'blue', width= 0.5) +
  insert_text(label = "rdpc")
}}
}
