% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{ipea_pal}
\alias{ipea_pal}
\title{Ipea palette}
\usage{
ipea_pal(
  palette = c("Blue", "Green", "Orange", "Pink", "Green-Blue", "Green-Blue-White",
    "Red-Blue", "Red-Blue-White", "Orange-Blue", "Orange-Blue-White", "Viridis",
    "Inferno", "Magma", "Plasma", "Cividis"),
  alpha = 1,
  begin = 0,
  end = 1,
  palette_direction = 1
)
}
\arguments{
\item{palette}{A character string indicating the color map option to use.
These options are available:
\itemize{
\item{}{'Blue'}
\item{}{'Green'}
\item{}{'Orange'}
\item{}{'Pink'}
\item{}{'Red-Blue'}
\item{}{'Orange-Blue'}
\item{}{'Green-Blue'}
\item{}{'Viridis'}
\item{}{'Inferno'}
\item{}{'Magma'}
\item{}{'Plasma'}
\item{}{'Cividis'}
}}

\item{alpha}{The alpha transparency in a number between \code{0} and \code{1}.}

\item{begin}{The (corrected) hue in a number between \code{0} and \code{1} at which the color map begins.}

\item{end}{The (corrected) hue in a number between \code{0} and \code{1} at which the color map ends.}

\item{palette_direction}{Sets the order of colors in the scale. If 1, the default,
colors are ordered from darkest to lightest. If -1, the order of
colors is reversed.}
}
\description{
Ipea palette
}
\examples{
scales::show_col(ipea_pal()(10))
scales::show_col(ipea_pal(palette_direction = -1)(6))
scales::show_col(ipea_pal(begin = 0.2, end = 0.8)(4))
scales::show_col(ipea_pal(palette = "Green")(6))
}
