% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{scale_ipea}
\alias{scale_ipea}
\alias{ipea_palette}
\alias{scale_fill_continuous_ipea}
\alias{scale_colour_continuous_ipea}
\alias{scale_colour_discrete_ipea}
\alias{scale_fill_discrete_ipea}
\title{Ipea Color Palette (Discrete) and Scales}
\usage{
ipea_pal()

scale_fill_continuous_ipea(...)

scale_colour_continuous_ipea(...)

scale_colour_discrete_ipea(...)

scale_fill_discrete_ipea(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}

\item{axis}{logical, whether the plot should keed the axis line or not.}

\item{only.color}{logical, whether there is more than one factor in the plot.}
}
\description{
A 9-color Ipea palette.
}
