% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ipea.R
\name{scale_fill_ipea}
\alias{scale_fill_ipea}
\title{Scale fill IPEA}
\usage{
scale_fill_ipea(
  discrete = F,
  palette = c("Blue", "Green", "Orange", "Pink", "Red-Blue", "Orange-Blue"),
  direction = c("horizontal", "vertical"),
  palette_direction = 1,
  hline,
  vline,
  show.limits = T,
  pt_br = T,
  barheight = 2,
  barwidth = 50,
  ...
)
}
\arguments{

\item{discrete}{A logical value indicating whether to use discrete or
continuous scale bar. The default value is continuous.}

\item{palette}{A character vector specifying the available palette for the
color palette. The default palette are "crimson", "orpheu", "cartola",
"caqui", "post", "wrapper", "blue_red", "ipea1", "ipea2", "ipea3", and
"manual".#' low, mid, high: Colors to be used for the low, mid, and
high values of the gradient, respectively.  \cr These parameters are
used when the "manual" option is selected.}

\item{direction}{A character vector specifying the legend ordering. \cr The available palette are "horizontal" and "vertical".
The default value is "horizontal".}

\item{palette_direction}{A logical argument specifying the direction of the color gradient.}

\item{show.limits}{A logical value indicating whether to display the color
gradient limits. The default value is TRUE.}

\item{pt_br}{A logical value indicating whether to use Brazilian Portuguese
formatting for labels. \cr If TRUE, decimal marks are set to ",", and
big marks are set to ".". The default value is TRUE.}

\item{hline}{Adjustment of the horizontal position of the extreme line of the graph axis. The default option is not to change the position.}

\item{vline}{Adjustment of the vertical position of the extreme line of the graph axis. The default option is not to change the position.}

\item{barheight}{The height of the color gradient bar. This parameter is used
when the direction is set to "vertical". The default value is 2.}

\item{barwidth}{The width of the color gradient bar. This parameter is used
when the direction is set to "horizontal". The default value is 50.}

\item{...}{Additional arguments to be passed to the scale_fill_gradientn or
scale_color_gradientn function from the ggplot2 package.}

\item{colours}{A vector of colors to be used for the gradient. This parameter
is used when palette other than "manual" are selected.}
}
\value{
A list object be added to a ggplot object to change color pallete.
}
\description{
Generate a fill palette (continuous or discrete) in the
formatting of texts published by the Institute for Applied
Economic Research (IPEA)
}
