% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{ipea_palette}
\alias{ipea_palette}
\title{Ipea Color Palette and Scales}
\usage{
ipea_palette(
  n,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  palette = c("Blue", "Green", "Orange", "Pink", "Red-Blue", "Orange-Blue")
)
}
\arguments{
\item{n}{The number of colors (\eqn{\ge 1}) to be in the palette.}

\item{alpha}{The alpha transparency in a number between \code{0} and \code{1}.}

\item{begin}{The (corrected) hue in a number between \code{0} and \code{1} at which the
color map begins.}

\item{end}{The (corrected) hue in a number between \code{0} and \code{1} at which the
color map ends.}

\item{direction}{Sets the order of colors in the scale. If \code{1}, the default,
colors are ordered from darkest to lightest. If \code{-1}, the order of
colors is reversed.}

\item{option}{A character string indicating the color map option to use.
Eight options are available:
\itemize{
\item{}{'Blue'}
\item{}{'Green'}
\item{}{'Orange'}
\item{}{'Pink'}
\item{}{'Red-Blue'}
\item{}{'Orange-Blue}
}}
}
\description{
This function creates a vector of n equally spaced colors
along the selected color map.
}
\details{
A 9-color Ipea palette.
}
