% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{ipea_pal}
\alias{ipea_pal}
\alias{scale_colour_ipea}
\alias{scale_color_ipea}
\alias{scale_fill_ipea}
\title{Ipea Color Palette (Discrete) and Scales}
\usage{
ipea_pal()

scale_colour_ipea(...)

scale_color_ipea(...)

scale_fill_ipea(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}

\item{axis}{logical, whether the plot should keed the axis line or not.}

\item{only.color}{logical, whether there is more than one factor in the plot.}
}
\description{
A 9-color Ipea palette.
}
