% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_palette.R
\name{ipea_palette}
\alias{ipea_palette}
\title{Ipea Color Palette and Scales}
\usage{
ipea_palette(palette = c("ipeatd", "ipea2", "ipea3"))
}
\arguments{
\item{palette}{A character vector specifying color/fill palette.}
}
\description{
A 9-color Ipea palette.
}
