% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaplot.R
\docType{package}
\name{ipeaplot}
\alias{ipeaplot}
\alias{_PACKAGE}
\alias{ipeaplot-package}
\title{ipeaplot: ggplot Graphics in Ipea Standard}
\description{
Convenient functions to create ggplot graphics following the editorial
guidelines of the  Institute for Applied Economic Research (IPEA).
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://ipeagit.github.io/flightsbr/}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeadata-lab/ipeaplot}
  \item Report bugs at \url{https://github.com/ipeadata-lab/ipeaplot/issues}
}

}
\author{
\strong{Maintainer}: Pedro Jorge \email{pedro.alves@ipea.gov.br}

Authors:
\itemize{
  \item Pedro Cavalcanti Gonçalves Ferreira \email{pedro.concalves@ipea.gov.br}
  \item Daniel Lima \email{daniel.viegas@ipea.gov.br}
  \item Gustavo Coelho \email{gustavo.coelho@ipea.gov.br}
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID})
  \item Lucas Mation \email{lucas.mation@ipea.gov.br}
}

}
