% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeaplot.R
\docType{package}
\name{ipeaplot}
\alias{ipeaplot}
\alias{_PACKAGE}
\alias{ipeaplot-package}
\title{ipeaplot: ggplot Graphics in Ipea Standard}
\description{
Convenient functions to create ggplot graphics following the editorial
guidelines of the  Institute for Applied Economic Research - Ipea.
}
\section{Usage}{

Please check the vignettes and data documentation on the
\href{https://github.com/ipeadata-lab/ipeaplot}{website}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ipeadata-lab/ipeaplot}
  \item Report bugs at \url{https://github.com/ipeadata-lab/ipeaplot/issues}
}

}
\author{
\strong{Maintainer}: Pedro Ferreira \email{pedro.ferreira2@ipea.gov.br}

Authors:
\itemize{
  \item Pedro Jorge \email{pedro.alves@ipea.gov.br}
  \item Daniel Lima \email{daniel.viegas@ipea.gov.br}
  \item Gustavo Coelho \email{gustavo.coelho@ipea.gov.br}
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com}
  \item Lucas Mation \email{lucas.mation@ipea.gov.br}
}

Other contributors:
\itemize{
  \item Fabio Vaz \email{fabio.vaz@ipea.gov.br} [contributor]
  \item Ipea - Institue for Applied Economic Research [copyright holder, funder]
}

}
