% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_eps.R
\name{save_eps}
\alias{save_eps}
\title{Save the chart in Eps format}
\usage{
save_eps(gplot, file.name, ...)
}
\arguments{
\item{gplot}{ggplot which will be saved.}

\item{file.name}{Character. Name of the file which will be generated}
}
\description{
Convenient function to save charts in Eps format.
}
\concept{save}
