% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{save_eps}
\alias{save_eps}
\title{Save the chart in .Eps}
\usage{
save_eps(gplot, file.name, width = 5, height = 3)
}
\arguments{
\item{gplot}{ggplot which will be saved.}

\item{file.name}{name of the file which will be generated}

\item{width}{width of the picture}

\item{height}{height of the picture}
}
\description{
This function makes easy to save charts in Eps.
}
