% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ipea.R
\name{scale_color_ipea}
\alias{scale_color_ipea}
\title{Scale color IPEA}
\usage{
scale_color_ipea(
  discrete = F,
  palette = c("Blue", "Green", "Orange", "Pink", "Green-Blue", "Red-Blue", "Green-Blue",
    "Orange-Blue", "Viridis", "Inferno", "Magma", "Plasma", "Cividis"),
  palette_direction = 1,
  decimal.mark = ",",
  barheight,
  barwidth,
  title.hjust,
  label.hjust,
  ...
)
}
\arguments{
\item{discrete}{A logical value indicating whether to use discrete or
continuous scale bar. The default value is continuous.}

\item{palette}{A character vector specifying the available palette for the
color palette. The default palette are \code{"Blue"}, but we can also change
to \code{'Green'}, \code{'Orange'}, \code{'Pink'}, \code{'Red-Blue'}, \code{'Orange-Blue'}, \code{'Green-Blue'},
\code{'Viridis'}, \code{'Inferno'}, \code{'Magma'}, \code{'Plasma'}, \code{'Cividis'}.}

\item{palette_direction}{A logical argument specifying if the ordering of the colors
will follow the default of the palette (when the argument is 1) or if it will
have an inverted ordering (for cases where it is -1).}

\item{decimal.mark}{The character to be used to indicate the numeric decimal point and
Character used between every 3 digits to separate thousands.
By default, the function uses a comma \code{","}, following
the format used in Brazilian Portuguese.}

\item{barheight}{The height of the color gradient bar. This parameter is used
when the direction is set to "horizontal".}

\item{barwidth}{The width of the color gradient bar. This parameter is used
when the direction is set to "horizontal".}

\item{title.hjust}{A number specifying horizontal justification of the title text.}

\item{label.hjust}{A number specifying vertical justification of the title text.}

\item{...}{Additional arguments to be passed to the scale_fill_gradientn,
scale_color_gradientn, scale_fill_distiller or scale_color_distiller
function from the ggplot2 package.}
}
\value{
A list object be added to a ggplot object to change color pallete.
}
\description{
Generate a color palette (continuous or discrete) following the
editorial guidelines used by the Institute for Applied Economic
Research - Ipea.
}
