% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ipea.R
\name{theme_ipea}
\alias{theme_ipea}
\title{Ggplot theme for IPEA graphics}
\usage{
theme_ipea(
  axis = c("none", "half", "full"),
  geom = c("bar", "line", "point", "sf"),
  text = T,
  legend.position,
  direction = c("horizontal", "vertical"),
  colours,
  ...
)
}
\arguments{
\item{axis}{A character vector specifying the axis style. Valid options are "none" (no axis lines),
"half" (half-length axis lines), and "full" (full-length axis lines).}

\item{text}{Logical value indicating whether to show text elements. If TRUE, axis text and ticks
will be displayed in black; otherwise, they will be hidden.}

\item{legend.position}{A character vector specifying the position of the legend. Valid options are
"right" (default), "left", "top", and "bottom".}

\item{...}{Additional arguments to be passed to the \code{theme} function from the \code{ggplot2} package.}

\item{box_option}{An integer indicating the box style. Use 1 for a dark background with white text
(default), and any other value for a white background with black text.}
}
\value{
A custom theme for IPEA graphics.
}
\description{
This function defines a custom theme for IPEA graphics, which provides standardized formatting
options for axis lines, text, legends, and panel borders.
}
