% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ipea.R
\name{theme_ipea}
\alias{theme_ipea}
\title{Ggplot theme for Ipea charts and figures}
\usage{
theme_ipea(
  axis = c("none", "half", "full"),
  geom = c("bar", "line", "point", "sf", "pie"),
  text = T,
  legend.position,
  grid.adjust = c("horizontal", "vertical"),
  x,
  y,
  yend,
  xend,
  x_breaks,
  y_breaks,
  bar_adjust = F,
  angle,
  include_ticks = T,
  ...
)
}
\arguments{
\item{axis}{A character vector specifying the axis style. Valid options are
\code{"none"} (no axis lines), \code{"half"} (half-length axis lines), and
\code{"full"} (full-length axis lines).}

\item{geom}{A character vector specifying some particularities of geom plot.
Valid options are \code{"bar"},\code{"line"}, \code{"point"}, \code{"sf"}.}

\item{text}{Logical value indicating whether to show text elements. If \code{TRUE},
axis text and ticks will be displayed in black; otherwise, they will
be hidden.}

\item{legend.position}{A character vector specifying the position of the
legend. Valid options are \code{"right"} (default), \code{"left"}, \code{"top"}, and
\code{"bottom"}.}

\item{grid.adjust}{Argument that defines whether the graphic specifications
will be in the standard horizontal or vertical legend.}

\item{x}{Manually set the initial x-axis value}

\item{y}{Manually set the initial y-axis value}

\item{yend}{Manually set the y-axis threshold value}

\item{xend}{Manually set the x-axis threshold value}

\item{x_breaks}{Specify values break amounts on the x-axis}

\item{y_breaks}{Specify values break amounts on the y-axis}

\item{bar_adjust}{Specific x-axis tweak for bar charts that should stick to the border.}

\item{angle}{Arrange x-axis angle}

\item{include_ticks}{Option to remove ticks and adjust chart.}

\item{...}{Additional arguments to be passed to the \code{theme} function from the
\code{ggplot2} package.}
}
\value{
A custom theme for IPEA graphics.
}
\description{
Applies a custom theme for ggplot figures following the editorial
guidelines used by the Institute for Applied Economic Research - Ipea. The
function includes standardized formatting of options for axis lines, text,
}
